Create user controller -  backend optimization
